yahooHockeyAnalyzer/src/requirements.txt
azure-functions
yahoo-fantasy-api
requests
pandas
numpy
yahoo-oauth